requirements.txt
python-telegram-bot==20.7
openai==1.13.3
google-cloud-bigquery==3.17.2
python-dotenv==1.0.1
requests==2.31.0
pillow==10.2.0